Export useTimer hook to other file
